Support software flow control
Add a new `-p` / `--preserve` flag to textadept-curses. When specified,
don’t grab the XON and XOFF control characters from the TTY driver. This
allows textadept-curses to run effectively on serial terminals that use
XON/XOFF flow control.

The `-p` flag was chosen to align with nano, where it has the same
behavior.