Added view.width, view.height, and "resize" event.
Previously it was not possible to set the size of a view that was the
second child of a split to a fixed size because it was not possible
to know the size of its parent. This commit amends the view metatable
to allow Lua to get and set the size of individual Scintilla widgets.
It also adds a "resize" event that is called whenever a Scintilla
widget changes sizes and passes a pointer to the associated view table
to the handler function. This allows Lua user scripts to "pin" a view
to a specific size regardless of window and parent split size changes
by defining a handler function that resets the size of a view on change.