Refactored platform resize calls to emit event directly
I realized that instead of having a wrapper function just to get
access to lua_pushdoc, I could add a couple of fake "types" for
emit() to pick up on, so that we can just directly push SciObject
and document pointers directly onto the stack from emit. I feel like
this simplifies things. There's also some code in here where I was
goofing around with trying to get GTK to allow me to resize the
command entry to less than 15 pixels tall, but so far nothing is
working...